recidive jail check
Added code to check and display the recidive fail2ban chain (Hestia only).
Minor rework on code.

+++ scripts/cm-server-mon.sh
@@ -10,7 +10,7 @@ MAILTO='[CHANGE ME]'
 SUBJECT="SERVER STATUS: $(hostname -f)"
 
 # Check if mailx is installed and assign it's path to a variable
-MAILX="$(which mailx)"
+MAILX="$(command -v mailx)"
 
 # Define the log file / Same name as script file but with .log extension
 LOGFILE=$0.log
@@ -40,72 +40,48 @@ if [[ $MAILX == "" ]]
 fi
 
 # This will print space usage by each directory inside directory $DIR, and after MAILX will send email with SUBJECT to MAILTO
-	echo "##### DISK USAGE FOR $DIR #####" > "$LOGFILE"
-	du -sh ${DIR}/* | sort -hr >> "$LOGFILE" 2>&1
-	echo "" >> "$LOGFILE"
+	{ echo "##### DISK USAGE FOR $DIR #####" ; du -sh ${DIR}/* | sort -hr ; echo "" ; } >> "$LOGFILE" 2>&1
 
 # This will print inode usage for /home directory
-	echo "##### INODE USAGE #####" >> "$LOGFILE"
-	df -hi >> "$LOGFILE" 2>&1
-	echo "" >> "$LOGFILE"
+	{ echo "##### INODE USAGE #####" ; df -hi ; echo "" ; } >> "$LOGFILE" 2>&1
 
 # Check MailQueue
-	echo "##### CHECKING THE MAIL QUEUE #####" >> "$LOGFILE"
-	/usr/sbin/exim -bp | /usr/sbin/exiqsumm >> "$LOGFILE" 2>&1
+	{ echo "##### CHECKING THE MAIL QUEUE #####" ; /usr/sbin/exim -bp | /usr/sbin/exiqsumm ; } >> "$LOGFILE" 2>&1
 
 # Check free space
-	echo "##### CHECKING FREE SPACE #####" >> "$LOGFILE"
-	df -Tha --total >> "$LOGFILE" 2>&1
-	echo "" >> "$LOGFILE"
+	{ echo "##### CHECKING FREE SPACE #####" ; df -Tha --total ; echo "" ; } >> "$LOGFILE" 2>&1
 
 # Check free memory
-	echo "##### CHECKING FREE MEMORY #####" >> "$LOGFILE"
-	free -mt >> "$LOGFILE" 2>&1
-	echo "" >> "$LOGFILE"
+	{ echo "##### CHECKING FREE MEMORY #####" ; free -mt ; echo "" ; } >> "$LOGFILE" 2>&1
 
+# Checking freespace before rsync rsync
+	{ echo "##### CHECKING FREE SPACE ON REMOTE BACKUP DEVICE #####" ; ssh $BAKUSER@$BAKSERV -i $RSYNKEY "df -h" ; } >> "$LOGFILE" 2>&1
+	
 # Perform rsync
-        echo "##### CHECKING FREE SPACE ON REMOTE BACKUP DEVICE #####" >> "$LOGFILE"
-        ssh $BAKUSER@$BAKSERV -i $RSYNKEY "df -h" >> "$LOGFILE" 2>&1
-	echo "##### CHECKING RSYNC BACKUP #####" >> "$LOGFILE"
-	echo "" >> "$LOGFILE"
-	rsync -ahv --no-g -e "ssh -p 22 -i $RSYNKEY" $LOCABAK $BAKUSER@$BAKSERV:$REMOBAK/$BAKFOLD/"$(hostname -f)" >> "$LOGFILE" 2>&1
-	echo "" >> "$LOGFILE"
+	{ echo "" ; echo "##### CHECKING RSYNC BACKUP #####" ; echo "" ; rsync -ahv --no-g -e "ssh -p 22 -i $RSYNKEY" $LOCABAK $BAKUSER@$BAKSERV:$REMOBAK/$BAKFOLD/"$(hostname -f)" ; echo "" ; } >> "$LOGFILE" 2>&1
 
 # Check fail2ban
-	echo "##### CHECKING FAIL2BAN #####" >> "$LOGFILE"
-	/usr/sbin/service fail2ban status >> "$LOGFILE" 2>&1
-	echo "
-	##### CHECKING JAIL SSH #####" >> "$LOGFILE"
-	fail2ban-client status ssh-iptables >> "$LOGFILE" 2>&1
-	echo "
-	##### CHECKING JAIL DOVECOT #####" >> "$LOGFILE"
-	fail2ban-client status dovecot-iptables >> "$LOGFILE" 2>&1
-	echo "
-	##### CHECKING JAIL EXIM #####" >> "$LOGFILE"
-	fail2ban-client status exim-iptables >> "$LOGFILE" 2>&1
-	echo "
-	##### CHECKING JAIL MYSQL #####" >> "$LOGFILE"
-	fail2ban-client status mysqld-iptables >> "$LOGFILE" 2>&1
-	echo "
-	##### CHECKING JAIL VSFTPD #####" >> "$LOGFILE"
-	fail2ban-client status vsftpd-iptables >> "$LOGFILE" 2>&1
-	echo "
-	##### CHECKING JAIL VESTA #####" >> "$LOGFILE"
-	fail2ban-client status $CPNAME-iptables >> "$LOGFILE" 2>&1
-	echo "" >> "$LOGFILE"
+	{ echo "##### CHECKING FAIL2BAN #####" ; /usr/sbin/service fail2ban status ; echo "" ; } >> "$LOGFILE" 2>&1
+	{ echo "##### CHECKING JAIL SSH #####" ; fail2ban-client status ssh-iptables ; echo "" ; } >> "$LOGFILE" 2>&1
+	{ echo "##### CHECKING JAIL DOVECOT #####" ; fail2ban-client status dovecot-iptables ; echo "" ; } >> "$LOGFILE" 2>&1
+	{ echo "##### CHECKING JAIL EXIM #####" ; fail2ban-client status exim-iptables ; echo "" ; } >> "$LOGFILE" 2>&1
+	{ echo "##### CHECKING JAIL MYSQL #####" ; fail2ban-client status mysqld-iptables ; echo "" ; } >> "$LOGFILE" 2>&1
+	{ echo "##### CHECKING JAIL VSFTPD #####" ; fail2ban-client status vsftpd-iptables ; echo "" ; } >> "$LOGFILE" 2>&1
+	{ echo "##### CHECKING JAIL VESTA/HESTIA #####" ; fail2ban-client status $CPNAME-iptables ; echo "" ; } >> "$LOGFILE" 2>&1
+	if [ $CPNAME = "hestia" ]; then
+		{ echo "##### CHECKING JAIL RECIDIVE #####" ; fail2ban-client status recidive ; echo "" ; } >> "$LOGFILE" 2>&1
+	fi
 
 # Check the number of outgoing messages per user
-	echo "##### CHECKING NUMBER OF MESSAGES PER USER #####" >> "$LOGFILE"
-	grep '<=' /var/log/exim4/mainlog | awk '{print $5}' | grep \@ | sort | uniq -c | sort -nrk1  >> "$LOGFILE" 2>&1
-	echo "" >> "$LOGFILE"
+	{ echo "##### CHECKING NUMBER OF MESSAGES PER USER #####" ; grep '<=' /var/log/exim4/mainlog | awk '{print $5}' | grep \@ | sort | uniq -c | sort -nrk1 ; echo "" ; } >> "$LOGFILE" 2>&1
 
 # Check failures in DNS
-	echo "##### CHECKING DNS FAILURES #####" >> "$LOGFILE"
-	tail -c 8192 /var/log/syslog | grep denied >> "$LOGFILE" 2>&1
-	echo "" >> "$LOGFILE"
+	{ echo "##### CHECKING DNS FAILURES #####" ; tail -c 8192 /var/log/syslog | grep denied ; echo "" ; } >> "$LOGFILE" 2>&1
 
-# Clean PHP session files older than 24h
-        for d in /home/*; do /usr/bin/find "$d"/tmp/sess_* -mmin +1440 -delete; done &> /dev/null
+# Clean PHP session files older than 24h (Vesta only / Hestia has a cron job for that
+	if [ $CPNAME = "vesta" ]; then
+		for d in /home/*; do /usr/bin/find "$d"/tmp/sess_* -mmin +1440 -delete; done &> /dev/null
+	fi
 
 # Send email alert
 	$MAILX -r root -s "$SUBJECT" "$MAILTO" < "$LOGFILE" 2>&1
